backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-multipart
pdfplumber
python-dotenv
